requirements.txt
constructs>=10.0.0,<11.0.0
aws-cdk-lib==2.38.1
importlib-metadata
importlib-resources
zipp
boto3
